Fix(all): fixes store crash bugs
