+++ client/store/constants.js
@@ -1,0 +1,1 @@
+export const [LOADING, LOADED, ERROR] = ['LOADING', 'LOADED', 'ERROR']
